refactor(alignment_io): hard-clean collect boundary
