+++ SpliceGrapher/formats/alignment_io/api.py
@@ -2,10 +2,16 @@
 
 from __future__ import annotations
 
+import os
 import sys
+from collections.abc import Mapping
+from os import PathLike
 from typing import Literal, cast, overload
 
-from SpliceGrapher.formats.depth_io import DepthSource
+import numpy
+
+from SpliceGrapher.formats.depth_io import DepthSource, is_depths_file, read_depths
+from SpliceGrapher.formats.junction import parse_junction_record
 
 from . import collect as collect_ops
 from .depths import calculate_gene_depths
@@ -21,6 +27,93 @@ from .types import (
     ReadDataSource,
     ReferencePath,
 )
+
+
+def _depth_map_to_arrays(depths: Mapping[str, list[int] | numpy.ndarray]) -> DepthMap:
+    """Normalize depth maps to int32 NumPy arrays."""
+    return {
+        chrom: numpy.asarray(chrom_depths, dtype=numpy.int32)
+        for chrom, chrom_depths in depths.items()
+    }
+
+
+def _is_depths_source(source: ReadDataSource) -> bool:
+    """Return ``True`` when ``source`` should be handled as a depths file."""
+    if not isinstance(source, (str, PathLike)):
+        return False
+
+    source_path = os.fspath(source)
+    lower = source_path.lower()
+    if lower.endswith((".sam", ".bam", ".cram")):
+        return False
+
+    if os.path.isfile(source_path):
+        try:
+            with open(source_path, "rb") as stream:
+                magic = stream.read(4)
+        except OSError:
+            magic = b""
+        if magic in {b"BAM\x01", b"CRAM"}:
+            return False
+
+    return is_depths_file(source_path)
+
+
+@overload
+def _collect_depths_source_data(
+    source: DepthSource,
+    *,
+    alignments: Literal[True],
+    include_depths: bool = True,
+    junctions: bool,
+    maxpos: int,
+    minanchor: int,
+    minjct: int,
+    verbose: bool,
+) -> CollectResultWithAlignments: ...
+
+
+@overload
+def _collect_depths_source_data(
+    source: DepthSource,
+    *,
+    alignments: Literal[False] = False,
+    include_depths: bool = True,
+    junctions: bool,
+    maxpos: int,
+    minanchor: int,
+    minjct: int,
+    verbose: bool,
+) -> CollectResult: ...
+
+
+def _collect_depths_source_data(
+    source: DepthSource,
+    *,
+    alignments: bool = False,
+    include_depths: bool = True,
+    junctions: bool,
+    maxpos: int,
+    minanchor: int,
+    minjct: int,
+    verbose: bool,
+) -> CollectResult | CollectResultWithAlignments:
+    """Bridge depths-file inputs into the public alignment I/O contract."""
+    depth_map, jcts = read_depths(
+        source,
+        parse_junction=parse_junction_record,
+        maxpos=maxpos,
+        minanchor=minanchor,
+        minjct=minjct,
+        depths=include_depths,
+        junctions=junctions,
+        verbose=verbose,
+    )
+    normalized_depths = _depth_map_to_arrays(depth_map)
+    if alignments:
+        empty_alignments: AlignmentMap = {}
+        return normalized_depths, jcts, empty_alignments
+    return normalized_depths, jcts
 
 
 def read_alignment_spans(
@@ -51,8 +144,8 @@ def read_alignment_depths(
     reference_fasta: ReferencePath = None,
 ) -> DepthMap:
     """Return read depths indexed by chromosome for SAM/BAM/CRAM sources."""
-    if collect_ops._is_depths_source(source):
-        depths, _ = collect_ops._collect_depths_source_data(
+    if _is_depths_source(source):
+        depths, _ = _collect_depths_source_data(
             cast(DepthSource, source),
             alignments=False,
             include_depths=True,
@@ -117,8 +210,8 @@ def read_alignment_junctions(
     reference_fasta: ReferencePath = None,
 ) -> JunctionMap:
     """Return splice junctions indexed by chromosome."""
-    if collect_ops._is_depths_source(source):
-        _, jcts = collect_ops._collect_depths_source_data(
+    if _is_depths_source(source):
+        _, jcts = _collect_depths_source_data(
             cast(DepthSource, source),
             alignments=False,
             include_depths=False,
@@ -184,9 +277,9 @@ def collect_alignment_data(
     reference_fasta: ReferencePath = None,
 ) -> CollectResult | CollectResultWithAlignments:
     """Return depths and junctions (and optionally alignments) from alignment input."""
-    if collect_ops._is_depths_source(source):
+    if _is_depths_source(source):
         if include_alignments:
-            return collect_ops._collect_depths_source_data(
+            return _collect_depths_source_data(
                 cast(DepthSource, source),
                 alignments=True,
                 include_depths=True,
@@ -196,7 +289,7 @@ def collect_alignment_data(
                 junctions=junctions,
                 verbose=verbose,
             )
-        return collect_ops._collect_depths_source_data(
+        return _collect_depths_source_data(
             cast(DepthSource, source),
             alignments=False,
             include_depths=True,
